\name{last}
\alias{last}
\title{function to return the last component of a vector}
\description{
function to extract the last component of a vector
}
\usage{
last(v)
}
\arguments{
  \item{v}{vector}
}
\value{
  The last component of the vector is returned
}
\author{Ott Toomet \email{siim@obs.ee}}
\seealso{\code{\link{length}}}
\examples{
a <- 1:5
last(a) # 5
}
\keyword{utilities}

