\name{micEcon-internal}
\alias{aidsCoef}
\alias{aidsRestr}
\alias{aidsSystem}
\alias{micEconIndex}
\alias{print.aidsEst}
\alias{print.aidsTestConsist}
\alias{print.heckit}
\alias{print.snqProfitImposeConvexity}
\alias{print.summary.maxLik}
\alias{print.summary.tobit2}
\alias{checkNames}
\alias{snqProfitCoef}
\alias{snqProfitCoefNames}
\alias{snqProfitRestrict}
\alias{snqProfitSystem}
\alias{snqProfitTestCoef}
\alias{summary.tobit2}
\alias{translogMonoRestr}


\title{ Internal micEcon Functions }

\description{
   Internal micEcon Functions
}

\details{
   These are not to be called by the user (or in some cases are just
   waiting for proper documentation to be written :).
}

\keyword{ internal }
